requirements.txt
streamlit
pandas
scikit-learn==1.6.1
joblib==1.5.0
psycopg2-binary
sqlalchemy
python-dotenv
